Allow customising the message included before list of changes to Slack
Without this it would be more fiddly to change the message to make it clearer
why the list of commits was being sent to Slack.

With this change the message could for instance say "A new version of Foo
Service has been deployed, these are the changes" or something to that extend.

The link to GitHub was also moved below the list of commits. It reads a bit
better.