Fix name of the Slack template input
Using a space to separate the two words look like it would work better than a
hyphen ('-') as you can't have a hyphen in an environment variable, which is
necessary for testing this locally.

Spaces are however explicitly changed to underscores in the @actions/core
library so it's clear how that would work, and it's prettier than using an
underscore through out.