Be smarter when finding suitable ref for `since`
This takes `until` into account to make sure we don't point `since` at the same
commit as `until`. This is done by first looking up `until` in case it's set to
the default "HEAD" value and then checking the commit SHA returned from that
against the list of tags on the repository.

The benefit of this is that users in many cases won't need to manually fill out
`since` because it can be inferred from the previous tag which should be
established in a pretty sensible way.